[REFACTOR] - Refactor of navbar components
