+++ src/components/Navbar/index.js
@@ -1,1 +1,1 @@
-export { default } from './Navbar.component';
+export { default } from './Navbar.component'
