--- src/components/VideoCard/VideoCardStyles.jsx
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-const VideoCardStyles = styled.div`
-  
-`;
-
-export default VideoCardStyles;
